fix: move logger file out of logger folder.
